Extracted Suggestion into its own component
